fix: run WME ops recursively instead of queuing
Queued ops mess with aggregate rebuilding.

If there are two cascading rules, each containing an aggregate, and the
upstream aggregate is rebuilding its partitions and issuing retractions
of the old partitions and assertions of the new ones, the WME op queue
and the token state may diverge in the downstream aggregate, and it will
not correctly autodelete produced WMEs.

Here a and b are output states of the upstream aggregator, and a' and b'
are their derivative output states of the downstream aggregator.

QUEUE HEAD    QUEUE PUSH
(any)         -a
              +a
              +b
-----------
-a            -a'
-----------
+a            +a'
-----------
+b            +b', but should also be -a'
              !! this does not see the state a' yet, so it cannot be
                 cleaned up in the partition rebuild, and it will be
                 inconsistently asserted in the following queue loop